* master: (36 commits)
  TFTPFileSystem::HandleAsynchRequestNeedPass: do not ask password for anonymous logins, first try to use empty password
  fix 97c5f179aa6225bf1e050f3781feb37d69d881cd
  refactoring: mark output parameters, reorder parameters
  fix 37a155e7d5e2204db5d64ffa3254f60e532c0cab (bugfix: FTP: resume download did not work)
  WebDAVFileSystem.cpp: Unused pointer value
  WebDAVFileSystem.cpp: cleanup
  version info: update
  "Connection failed" dialog: make "Reconnect" button default
  fix code style
  TFarButton
  TQueryButtonAlias
  remove unnecessary brackets
  fix code style
  version info: update
  minor code style (catch (Exception & E))
  refactoring: rename EScpSkipFile --> ESkipFile; EScpFileSkipped --> EFileSkipped
  refactoring: use WEBDAV_NO_ERROR instead of 0
  refactoring: WEBDAV_ERR_ASSERT
  refactoring: rename function
  M#2602 Processing of copy file don't break
  ...

Conflicts:
	src/NetBox/FarDialog.cpp
	src/NetBox/FarPlugin.cpp
	src/NetBox/WinSCPFileSystem.cpp

+++ src/core/RemoteFiles.h
@@ -147,7 +147,7 @@ private:
   TModificationFmt FModificationFmt;
   int64_t FSize;
   UnicodeString FFileName;
-  Integer FINodeBlocks;
+  int64_t FINodeBlocks;
   TDateTime FModification;
   TDateTime FLastAccess;
   TRemoteToken FGroup;
